+++ src/Installers/InertiaInstaller.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace TPG\Vitamin\Installers;
 
+use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Str;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Process\Process;
@@ -27,10 +30,7 @@ class InertiaInstaller implements InstallerContract
         $this->output->write('.');
 
         $this->installMiddleware();
-        $this->output->write('.');
-
         $this->buildRoutes();
-        $this->output->write('.');
 
         $this->output->writeln('[<info>✔</info>]');
     }
@@ -39,11 +39,23 @@ class InertiaInstaller implements InstallerContract
     {
         $process = Process::fromShellCommandline('php ./artisan inertia:middleware');
         $process->run();
+        $this->output->write('.');
+
+        $lines = collect(File::lines(app_path('Http/Kernel.php'))->toArray());
+        $index = $lines->filter(fn ($line) => strstr($line, 'SubstituteBindings::class'))->keys()->first();
+        $after = $lines->splice($index + 1);
+
+        $lines->push(str_repeat(' ', 8).'App\Middleware\HandleInertiaRequests::class,');
+        $lines = $lines->concat($after);
+
+        file_put_contents(app_path('Http/Kernel.php'), implode("\n", $lines->toArray()));
+        $this->output->write('.');
     }
 
     protected function buildRoutes(): void
     {
         $process = Process::fromShellCommandline('yarn routes');
         $process->run();
+        $this->output->write('.');
     }
 }