+++ stubs/AppComposer.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Http\Composers;
+
+class AppComposer extends TPG\Vitamin\Composers\AppComposer
+{
+
+}