+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.11] 2021-08-20
+### Changed
+- The init command will no longer ask for a "components" path.
+
+### Added
+- The Inertia installer will add the `HandleInertiaRequests` middleware to the `Kernel.php` file.
+- Removed the execution timeout from a number of processes.
+
 ## [0.0.10] 2021-08-18
 ### Fixed
 - Fixed a bug not installing node dependencies correctly.